tests/wpt/web-platform-tests/html/semantics/embedded-content/the-iframe-element/move_iframe_in_dom_04.html

<!DOCTYPE html>
<title>moving modified IFRAME in document (original page from server, document.write modification)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/#iframe-load-event-steps">
<iframe src="support/blank.htm"></iframe>
<div id="target"></div>
<script>
setup({ single_test: true });
onload = function(){
  var ifr = document.getElementsByTagName('iframe')[0];
  ifr.contentDocument.open();
  ifr.contentDocument.write('Modified document');
  ifr.contentDocument.close();
  setTimeout(function() {
    ifr.onload = function () {
      assert_equals(ifr.contentDocument.body.textContent.indexOf('Modified'), -1);
      done();
    };
    document.getElementById('target').appendChild(ifr);
  }, 100);
}
</script>
